# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_JanArsMAI_Trafic-Incident-Service_git_internal_presentation_dto.AddDriverDto:
    properties:
      date_birth:
        type: string
      experience:
        type: integer
      license_issue_date:
        type: string
      license_number:
        type: string
      name:
        type: string
    type: object
...
  github_com_JanArsMAI_Trafic-Incident-Service_git_internal_presentation_dto.AddVehicleDto:
    properties:
      model:
        type: string
      number:
        type: string
      owner:
        type: integer
      type:
        type: string
      year:
        type: integer
    type: object
...
  github_com_JanArsMAI_Trafic-Incident-Service_git_internal_presentation_dto.DriverResponse:
    properties:
      accidents:
        type: integer
      birthdate:
        type: string
      created_at:
        type: string
      experience:
        type: integer
      full_name:
        type: string
      id:
        type: integer
      license_exp_date:
        type: string
      license_id:
        type: string
    type: object
  github_com_JanArsMAI_Trafic-Incident-Service_git_internal_presentation_dto.DriversResponse:
    properties:
      drivers:
        items:
          $ref: '#/definitions/github_com_JanArsMAI_Trafic-Incident-Service_git_internal_presentation_dto.DriverResponse'
        type: array
    type: object
  github_com_JanArsMAI_Trafic-Incident-Service_git_internal_presentation_dto.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  github_com_JanArsMAI_Trafic-Incident-Service_git_internal_presentation_dto.UpdateDriverDto:
    properties:
      date_birth:
        type: string
      experience:
        type: integer
      license:
        type: string
      license_issue_date:
        type: string
      name:
        type: string
      new_license:
        type: string
    type: object
...
  github_com_JanArsMAI_Trafic-Incident-Service_git_internal_presentation_dto.UpdateVehicleDto:
    properties:
      model:
        type: string
      number:
        type: string
      owner:
        type: integer
      type:
        type: string
      year:
        type: integer
    type: object
...
  github_com_JanArsMAI_Trafic-Incident-Service_git_internal_presentation_dto.VehicleResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      model:
        type: string
      number:
        type: string
      owner:
        type: integer
      type:
        type: string
      year:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API для управления пользователями.
  title: Traffic Incident Service API
  version: "1.0"
paths:
...
  /drivers/add:
    post:
      consumes:
      - application/json
      description: Создает нового водителя по переданным данным
      parameters:
      - description: Данные водителя
        in: body
        name: driver
        required: true
        schema:
          $ref: '#/definitions/github_com_JanArsMAI_Trafic-Incident-Service_git_internal_presentation_dto.AddDriverDto'
      produces:
      - application/json
      responses:
        "201":
          description: 'driver_id: ID созданного водителя'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Некорректные данные или неверный формат даты
          schema:
            $ref: '#/definitions/github_com_JanArsMAI_Trafic-Incident-Service_git_internal_presentation_dto.ErrorResponse'
        "409":
          description: Водитель с таким номером лицензии уже существует
          schema:
            $ref: '#/definitions/github_com_JanArsMAI_Trafic-Incident-Service_git_internal_presentation_dto.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/github_com_JanArsMAI_Trafic-Incident-Service_git_internal_presentation_dto.ErrorResponse'
      summary: Добавить нового водителя
      tags:
      - Водители
  /drivers/get_by_license/{license}:
    get:
      consumes:
      - application/json
      description: Возвращает данные водителя
      parameters:
      - description: Номер лицензии
        in: path
        name: license
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_JanArsMAI_Trafic-Incident-Service_git_internal_presentation_dto.DriverResponse'
        "400":
          description: Пустой license
          schema:
            $ref: '#/definitions/github_com_JanArsMAI_Trafic-Incident-Service_git_internal_presentation_dto.ErrorResponse'
        "404":
          description: Водитель не найден
          schema:
            $ref: '#/definitions/github_com_JanArsMAI_Trafic-Incident-Service_git_internal_presentation_dto.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/github_com_JanArsMAI_Trafic-Incident-Service_git_internal_presentation_dto.ErrorResponse'
      summary: Получить водителя по номеру лицензии
      tags:
      - Водители
  /drivers/get_by_name/{name}:
    get:
      consumes:
      - application/json
      description: Возвращает список всех водителей с указанным именем (частичное
        совпадение допускается)
      parameters:
      - description: Имя водителя
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_JanArsMAI_Trafic-Incident-Service_git_internal_presentation_dto.DriversResponse'
        "400":
          description: Пустое имя
          schema:
            $ref: '#/definitions/github_com_JanArsMAI_Trafic-Incident-Service_git_internal_presentation_dto.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/github_com_JanArsMAI_Trafic-Incident-Service_git_internal_presentation_dto.ErrorResponse'
      summary: Получить список водителей по имени
      tags:
      - Водители
...
  /drivers/update:
    patch:
      consumes:
      - application/json
      description: Частичное обновление данных водителя по номеру лицензии
      parameters:
      - description: Данные для обновления
        in: body
        name: driver
        schema:
          $ref: '#/definitions/github_com_JanArsMAI_Trafic-Incident-Service_git_internal_presentation_dto.UpdateDriverDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/github_com_JanArsMAI_Trafic-Incident-Service_git_internal_presentation_dto.ErrorResponse'
        "404":
          description: Водитель не найден
          schema:
            $ref: '#/definitions/github_com_JanArsMAI_Trafic-Incident-Service_git_internal_presentation_dto.ErrorResponse'
        "409":
          description: Новый номер лицензии уже используется
          schema:
            $ref: '#/definitions/github_com_JanArsMAI_Trafic-Incident-Service_git_internal_presentation_dto.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/github_com_JanArsMAI_Trafic-Incident-Service_git_internal_presentation_dto.ErrorResponse'
      summary: Обновить данные водителя
      tags:
      - Водители
...
  /vehicles/{number}:
    get:
      consumes:
      - application/json
      description: Возвращает информацию о транспортном средстве по его государственному
        номеру.
      parameters:
      - description: номер ТС
        in: path
        name: number
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Информация о транспортном средстве
          schema:
            $ref: '#/definitions/github_com_JanArsMAI_Trafic-Incident-Service_git_internal_presentation_dto.VehicleResponse'
        "400":
          description: Пустой номер
          schema:
            $ref: '#/definitions/github_com_JanArsMAI_Trafic-Incident-Service_git_internal_presentation_dto.ErrorResponse'
        "404":
          description: ТС не найдено
          schema:
            $ref: '#/definitions/github_com_JanArsMAI_Trafic-Incident-Service_git_internal_presentation_dto.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/github_com_JanArsMAI_Trafic-Incident-Service_git_internal_presentation_dto.ErrorResponse'
      summary: Получить транспортное средство по номеру
      tags:
      - Транспорт
  /vehicles/add:
    post:
      consumes:
      - application/json
      description: Создаёт новое транспортное средство. Номер должен быть уникальным.
      parameters:
      - description: Данные нового транспорта
        in: body
        name: data
        schema:
          $ref: '#/definitions/github_com_JanArsMAI_Trafic-Incident-Service_git_internal_presentation_dto.AddVehicleDto'
      produces:
      - application/json
      responses:
        "201":
          description: ID созданного ТС
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Неверные данные
          schema:
            $ref: '#/definitions/github_com_JanArsMAI_Trafic-Incident-Service_git_internal_presentation_dto.ErrorResponse'
        "409":
          description: ТС с таким номером уже существует
          schema:
            $ref: '#/definitions/github_com_JanArsMAI_Trafic-Incident-Service_git_internal_presentation_dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_JanArsMAI_Trafic-Incident-Service_git_internal_presentation_dto.ErrorResponse'
      summary: Добавить транспортное средство
      tags:
      - Транспорт
  /vehicles/update:
    patch:
      consumes:
      - application/json
      description: Обновляет переданные поля транспортного средства по его номеру.
      parameters:
      - description: Данные для обновления ТС
        in: body
        name: data
        schema:
          $ref: '#/definitions/github_com_JanArsMAI_Trafic-Incident-Service_git_internal_presentation_dto.UpdateVehicleDto'
      produces:
      - application/json
      responses:
        "200":
          description: Сообщение об успешном обновлении
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверные данные
          schema:
            $ref: '#/definitions/github_com_JanArsMAI_Trafic-Incident-Service_git_internal_presentation_dto.ErrorResponse'
        "404":
          description: ТС не найдено
          schema:
            $ref: '#/definitions/github_com_JanArsMAI_Trafic-Incident-Service_git_internal_presentation_dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка
          schema:
            $ref: '#/definitions/github_com_JanArsMAI_Trafic-Incident-Service_git_internal_presentation_dto.ErrorResponse'
      summary: Обновить данные транспортного средства
      tags:
      - Транспорт
securityDefinitions:
  BearerAuth:
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"